Refactor package structure and moved commands to individual classes
