+++ src/main/java/Commands/Command.java
@@ -1,3 +1,5 @@
+package Commands;
+
 @FunctionalInterface
 public interface Command {
     void execute();
