+++ src/main/java/Commands/FlightCommand/HoldCommand.java
@@ -1,0 +1,17 @@
+package Commands.FlightCommand;
+
+import Commands.Command;
+import Flight.Flight;
+
+public class HoldCommand implements Command {
+    private final Flight flight;
+
+    public HoldCommand(Flight flight) {
+        this.flight = flight;
+    }
+
+    @Override
+    public void execute() {
+        flight.hold();
+    }
+}
