+++ src/main/java/Commands/GameCommand/ControlFlightCommand.java
@@ -1,0 +1,69 @@
+package Commands.GameCommand;
+
+import Commands.Command;
+import Flight.Flight;
+import Interceptors.InterceptorDispatcher;
+import Commands.FlightCommand.*;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Scanner;
+
+public class ControlFlightCommand implements Command {
+    private final List<Flight> flights;
+    private final Scanner scanner;
+    private final InterceptorDispatcher dispatcher;
+    private static final Map<String, Command> commands = new HashMap<>();
+
+    public ControlFlightCommand(List<Flight> flights, Scanner scanner, InterceptorDispatcher dispatcher) {
+        this.flights = flights;
+        this.scanner = scanner;
+        this.dispatcher = dispatcher;
+    }
+
+    @Override
+    public void execute() {
+        if (flights.isEmpty()) {
+            System.out.println("No flights available!");
+            return;
+        }
+
+        new ListFlightsCommand(flights).execute();
+        System.out.print("Enter flight number to control: ");
+        String flightNumber = scanner.nextLine().toUpperCase();
+
+        Flight selectedFlight = flights.stream()
+                .filter(f -> f.getFlightNumber().equals(flightNumber))
+                .findFirst()
+                .orElse(null);
+
+        if (selectedFlight == null) {
+            System.out.println("Flight.Flight not found!");
+            return;
+        }
+
+        System.out.println("\n=== Flight.Flight Controls ===");
+        System.out.println("1. Take off");
+        System.out.println("2. Land");
+        System.out.println("3. Hold");
+        System.out.print("Choose action: ");
+
+        String action = scanner.nextLine();
+        dispatcher.dispatch("Controlling " + flightNumber + action);
+
+        setUpFlightCommands(selectedFlight);
+        Command command = commands.get(action);
+        if (command != null) {
+            command.execute();
+        } else {
+            System.out.println("Invalid option!");
+        }
+    }
+
+    private static void setUpFlightCommands(Flight selectedFlight) {
+        commands.put("1", new TakeOffCommand(selectedFlight));
+        commands.put("2", new LandCommand(selectedFlight));
+        commands.put("3", new HoldCommand(selectedFlight));
+    }
+}
