+++ src/main/java/Flight/FlightTypes/FlightType.java
@@ -1,3 +1,5 @@
+package Flight.FlightTypes;
+
 public enum FlightType {
     PRIVATE,
     PASSENGER,
