+++ src/main/java/Interceptors/Interceptor.java
@@ -1,3 +1,5 @@
+package Interceptors;
+
 public interface Interceptor {
     void handleRequest(String input);
 }
