+++ src/main/java/Interceptors/InterceptorDispatcher.java
@@ -1,20 +1,19 @@
+package Interceptors;
+
 import java.util.List;
 import java.util.ArrayList;
 
-public class InterceptorDispatcher{
+public class InterceptorDispatcher {
     // manages interceptors
     private final List<Interceptor> interceptorList = new ArrayList<>();
 
-    public void addInterceptor(Interceptor i){
+    public void addInterceptor(Interceptor i) {
         interceptorList.add(i);
     }
 
-    public void dispatch(String input){
-        for(Interceptor i : interceptorList){
+    public void dispatch(String input) {
+        for (Interceptor i : interceptorList) {
             i.handleRequest(input);
         }
-
     }
-
-
 }
