+++ src/main/java/Simulator.java
@@ -1,3 +1,9 @@
+import Commands.Command;
+import Commands.GameCommand.*;
+import Flight.Flight;
+import Interceptors.*;
+import WeatherPubSub.WeatherStation;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -9,7 +15,7 @@ public class Simulator {
     private static List<Flight> flights;
     private static Scanner scanner;
     private static final InterceptorDispatcher dispatcher = new InterceptorDispatcher();
-    private static final Map<String, GameCommand> commands = new HashMap<>();
+    private static final Map<String, Command> commands = new HashMap<>();
 
     public static void startSimulation() {
         initialise();
@@ -32,7 +38,7 @@ public class Simulator {
             String choice = scanner.nextLine().toUpperCase();
             dispatcher.dispatch(choice);
 
-            GameCommand command = commands.get(choice);
+            Command command = commands.get(choice);
             if (command != null) {
                 command.execute();
                 if (command instanceof QuitCommand) {
