+++ src/main/java/WeatherPubSub/Publisher.java
@@ -1,3 +1,5 @@
+package WeatherPubSub;
+
 public interface Publisher {
     void publish(String topic, String message);
 }
