+++ src/main/java/WeatherPubSub/Subscriber.java
@@ -1,3 +1,5 @@
+package WeatherPubSub;
+
 public interface Subscriber {
     void receive(String topic, String message);
 }
