+++ src/test/java/AutomatedTest.java
@@ -1,3 +1,7 @@
+import Decorators.FuelDecorator;
+import Flight.Flight;
+import Flight.FlightFactory;
+import Flight.FlightTypes.FlightType;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Assertions;
@@ -15,13 +19,13 @@ public class AutomatedTest {
     @Test
     void testFlightCreation() {
         // Test passenger flight
-        Assertions.assertEquals("PassengerFlight", passengerFlight.getType());
+        Assertions.assertEquals("Flight.FlightTypes.PassengerFlight", passengerFlight.getType());
         Assertions.assertEquals("PA123", passengerFlight.getFlightNumber());
         Assertions.assertEquals("On ground/Runway", passengerFlight.getState());
         Assertions.assertEquals(100, passengerFlight.getFuel());
 
         // Test military flight
-        Assertions.assertEquals("MilitaryFlight", militaryFlight.getType());
+        Assertions.assertEquals("Flight.FlightTypes.MilitaryFlight", militaryFlight.getType());
         Assertions.assertEquals("MIL456", militaryFlight.getFlightNumber());
     }
 
@@ -80,8 +84,8 @@ public class AutomatedTest {
         Flight cargoFlight = FlightFactory.createFlight(FlightType.CARGO, "CG789");
         Flight privateFlight = FlightFactory.createFlight(FlightType.PRIVATE, "PV101");
         
-        Assertions.assertEquals("CargoFlight", cargoFlight.getType());
-        Assertions.assertEquals("Private Flight", privateFlight.getType());
+        Assertions.assertEquals("Flight.FlightTypes.CargoFlight", cargoFlight.getType());
+        Assertions.assertEquals("Private Flight.Flight", privateFlight.getType());
     }
 
     @Test
